relocate csv logger to sioclient
